JS: split CodeInjection.qll into three parts
Splits `CodeInjection.qll` into three parts, primarily to allow for
avoiding module activation of file modules that contain
`Configuration` classes. This allows us to extend the sinks and
sources of a `Configuration` without causing a serious compilation
speed or query speed degradation. That is the motivation for
introducing `CodeInjection/Configration.qll` and
`CodeInjection/Interface.qll`.

The related, but less important, introduction of
`CodeInjection/Implementation.qll` allows reusing the `Configuration`
in queries with a specialized set of sinks and sources.

The `CodeInjection/Implementation.qll` design is not yet optimal,
since it introduces sources and sinks when it is imported. This makes
it slightly tricky to make a query that reuses all of the default
sources, but none of the default sinks (this design is useful to avoid
double flagging with similar queries). If we find the need, we can
introduce yet another layer of abstract classes to support that use
case, the current design does not prevent us from doing that later.